pointer address

+++ print_XS.c
@@ -23,7 +23,7 @@ int print_XS(unsigned int num)
 		counter++;
 	}
 	counter++;
-	hex = malloc(sizeof(int) * counter);
+	hex = malloc(sizeof(long int) * counter);
 
 	for (i = 0; i < counter; i++)
 	{
